+++ elfpy/eth/accounts/__init__.py
@@ -1,3 +1,3 @@
 """Helper utilities for creating and managing accounts"""
 
-from .agent_account import EthAccount
+from .eth_account import EthAccount
